+++ sfsimodels/__init__.py
@@ -2,7 +2,8 @@ from sfsimodels.models.abstract_models import PhysicalObject, CustomObject
 from sfsimodels.models.hazards import SeismicHazard
 from sfsimodels.models.foundations import Foundation, PadFoundation, RaftFoundation
 from sfsimodels.models.soils import Soil, CriticalSoil, discretize_soil_profile, SoilProfile, StressDependentSoil
-from sfsimodels.models.buildings import Building, FrameBuilding, WallBuilding, SDOFBuilding, FrameBuilding2D, Section
+from sfsimodels.models.buildings import Building, FrameBuilding, WallBuilding, SDOFBuilding, FrameBuilding2D, \
+    Section, NullBuilding
 
 # deprecated
 from sfsimodels.models.soils import SoilStressDependent, SoilCritical