+++ sfsimodels/models/__init__.py
@@ -1,7 +1,8 @@
 from sfsimodels.models.hazards import SeismicHazard
 from sfsimodels.models.foundations import Foundation, FoundationPad, FoundationRaft
 from sfsimodels.models.soils import Soil, CriticalSoil, StressDependentSoil, SoilCritical, SoilProfile
-from sfsimodels.models.buildings import Building, BuildingFrame, BuildingWall, BuildingSDOF, BuildingFrame2D, Section, SDOFBuilding
+from sfsimodels.models.buildings import Building, BuildingFrame, BuildingWall, BuildingSDOF, BuildingFrame2D, \
+    Section, SDOFBuilding, NullBuilding
 from sfsimodels.models import material
 from sfsimodels.models.systems import SoilStructureSystem